+++ db/updates/update_6.5-6.6.sql
@@ -1,1 +1,0 @@
-ALTER TABLE `formulasMetaData` ADD `src` INT NOT NULL DEFAULT '0' COMMENT '0 = pvLocal, 1 = pvMarket' AFTER `rating`;