+++ releasenotes.md
@@ -1,4 +1,11 @@
-Whats New in v9.7
+Whats New in v9.8
 --------------------------
-- Bug fix release
+- Added phpMyAdmin definition in the docker compose
+- Added PV Url in settings, you have to update it according to your current setup
+- Added config / integration for the PV scale
+- Scheduled formulas improvements
+- Added a low stock icon in formula view
+- Added an automated backup to Google Drive solution - Currently not publicly available
+- Info message replaced with a toast style message
+- IFRA Lib under the hood db improvements
 - For full details please refer to the CHANGELOG